refactor completed?
namespace is so much cleaner.

added module level docstrings

docstrings for most functions added

all unit tests passed
