+++ docs/exceptions.rst
@@ -1,14 +1,14 @@
 Exceptions
 -------------
 
-.. autoclass:: markdown2canvas.AlreadyExists
+.. autoclass:: markdown2canvas.exception.AlreadyExists
    :members:
    :undoc-members:
 
-.. autoclass:: markdown2canvas.SetupError
+.. autoclass:: markdown2canvas.exception.SetupError
    :members:
    :undoc-members:
 
-.. autoclass:: markdown2canvas.DoesntExist
+.. autoclass:: markdown2canvas.exception.DoesntExist
    :members:
    :undoc-members:
