+++ docs/free_functions.rst
@@ -1,3 +1,30 @@
 Free functions
-------------------
+=====================
+
+
+Setup functions
+-------------------
+
+.. automodule:: markdown2canvas.setup_functions
+   :members:
+
+
+
+Functions for interacting with a course on Canvas
+---------------------------------------------------
+
+.. automodule:: markdown2canvas.course_interaction_functions
+   :members:
+
+
+Functions markdown2canvas uses to translate from markdown to Canvas-html
+--------------------------------------------------------------------------
+
+.. automodule:: markdown2canvas.translation_functions
+   :members:
+
+.. autofunction:: markdown2canvas.canvas_objects.find_local_images
+
+.. autofunction:: markdown2canvas.canvas_objects.find_local_files
+
