+++ docs/requirements.txt
@@ -1,2 +1,3 @@
 canvasapi
 Pygments
+markdown
