+++ test/test_assignment.py
@@ -64,7 +64,7 @@ class TestAssignment():
 		assignment.publish(course,overwrite=True)
 
 		# the second publish, with overwrite=False, should raise
-		with pytest.raises(mc.AlreadyExists):
+		with pytest.raises(mc.exception.AlreadyExists):
 			assignment.publish(course,overwrite=False) # default is False
 
 	def test_doesnt_find_deleted(self, course, assignment):
