+++ test/test_download_pages_to_markdown.py
@@ -32,7 +32,7 @@ class DownloadTester(unittest.TestCase):
 		if os.path.exists(destination):
 			shutil.rmtree(destination)
 
-		mc.download_pages(destination, self.course, even_if_exists=False)
+		mc.canvas2markdown.download_pages(destination, self.course, even_if_exists=False)
 
 	def test_aaa_can_download_some_pages(self):
 		import os, shutil
@@ -41,7 +41,7 @@ class DownloadTester(unittest.TestCase):
 			shutil.rmtree(destination)
 
 		my_filter = lambda title: 'test' in title.lower()
-		mc.download_pages(destination, self.course, even_if_exists=False, name_filter=my_filter)
+		mc.canvas2markdown.download_pages(destination, self.course, even_if_exists=False, name_filter=my_filter)
 
 
 if __name__ == '__main__':
