+++ test/test_link.py
@@ -51,7 +51,7 @@ class TestLink():
 		link.publish(course,overwrite=True)
 
 		# the second publish, with overwrite=False, should raise
-		with pytest.raises(mc.AlreadyExists):
+		with pytest.raises(mc.exception.AlreadyExists):
 			link.publish(course,overwrite=False) # default is False
