+++ test/test_style.py
@@ -39,7 +39,7 @@ class TestStyle():
 		page_uses_droplets_via_style.publish(course,overwrite=True)
 
 		# the second publish, with overwrite=False, should raise
-		with pytest.raises(mc.AlreadyExists):
+		with pytest.raises(mc.exception.AlreadyExists):
 			page_uses_droplets_via_style.publish(course,overwrite=False) # default is False
